// examples/order-flow/server/src/language-server/generated/module.ts
/******************************************************************************
 * This file was generated by langium-cli 4.3.0.
 * DO NOT EDIT MANUALLY!
 ******************************************************************************/

import type { LangiumSharedCoreServices, LangiumCoreServices, LangiumGeneratedCoreServices, LangiumGeneratedSharedCoreServices, LanguageMetaData, Module } from 'langium';
import { OrderFlowAstReflection } from './ast.js';
import { DomainGrammar, LayoutGrammar, ProcessGrammar } from './grammar.js';

export const DomainLanguageMetaData = {
    languageId: 'order-flow-domain',
    fileExtensions: ['.domain'],
    caseInsensitive: false,
    mode: 'development'
} as const satisfies LanguageMetaData;

export const LayoutLanguageMetaData = {
    languageId: 'order-flow-layout',
    fileExtensions: ['.layout'],
    caseInsensitive: false,
    mode: 'development'
} as const satisfies LanguageMetaData;

export const ProcessLanguageMetaData = {
    languageId: 'order-flow-process',
    fileExtensions: ['.process'],
    caseInsensitive: false,
    mode: 'development'
} as const satisfies LanguageMetaData;

export const OrderFlowGeneratedSharedModule: Module<LangiumSharedCoreServices, LangiumGeneratedSharedCoreServices> = {
    AstReflection: () => new OrderFlowAstReflection()
};

export const DomainGeneratedModule: Module<LangiumCoreServices, LangiumGeneratedCoreServices> = {
    Grammar: () => DomainGrammar(),
    LanguageMetaData: () => DomainLanguageMetaData,
    parser: {}
};

export const LayoutGeneratedModule: Module<LangiumCoreServices, LangiumGeneratedCoreServices> = {
    Grammar: () => LayoutGrammar(),
    LanguageMetaData: () => LayoutLanguageMetaData,
    parser: {}
};

export const ProcessGeneratedModule: Module<LangiumCoreServices, LangiumGeneratedCoreServices> = {
    Grammar: () => ProcessGrammar(),
    LanguageMetaData: () => ProcessLanguageMetaData,
    parser: {}
};